todo/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
numpy==1.21.3
openpyxl==3.0.9
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
pandas==1.3.4
promise==2.3
psycopg2==2.9.3
Pygments==2.11.2
pyparsing==3.0.7
pyquery==1.4.3
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
